requirements.txt
fastapi~=0.103.2
uvicorn
python-multipart
requests
sqlalchemy~=2.0.21
jinja2
pydantic~=2.4.2
starlette~=0.27.0
aioredis
requests